# cms-server/docs/swagger.yaml
basePath: /
definitions:
...
  business.CmsAdSeat:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      ID:
        description: 主键ID
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      desc:
        type: string
      height:
        type: integer
      mapData:
        additionalProperties:
          type: string
        description: key-val 模式数据
        type: object
      name:
        type: string
      status:
        type: integer
      style:
        type: string
      width:
        type: integer
    type: object
...
  business.CmsCat:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      ID:
        description: 主键ID
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      alias:
        type: string
      beNav:
        type: boolean
      beSys:
        type: boolean
      desc:
        type: string
      groupId:
        type: integer
      keywords:
        type: string
      mapData:
        additionalProperties:
          type: string
        description: key-val 模式数据
        type: object
      mediaType:
        type: integer
      name:
        type: string
      pid:
        type: integer
      sort:
        type: integer
      status:
        type: integer
      thumb:
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /cmsAdSeat/createCmsAdSeat:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建CmsAdSeat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsAdSeat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建CmsAdSeat
      tags:
      - CmsAdSeat
  /cmsAdSeat/deleteCmsAdSeat:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除CmsAdSeat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsAdSeat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除CmsAdSeat
      tags:
      - CmsAdSeat
  /cmsAdSeat/deleteCmsAdSeatByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除CmsAdSeat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除CmsAdSeat
      tags:
      - CmsAdSeat
  /cmsAdSeat/findCmsAdSeat:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: desc
        type: string
      - in: query
        name: height
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: status
        type: integer
      - in: query
        name: style
        type: string
      - in: query
        name: width
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询CmsAdSeat
      tags:
      - CmsAdSeat
  /cmsAdSeat/getCmsAdSeatList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: desc
        type: string
      - in: query
        name: height
        type: integer
      - in: query
        name: name
        type: string
      - description: 排序方式:升序false(默认)|降序true
        in: query
        name: orderDesc
        type: boolean
      - description: Add By ljd 20210708, 统一在这里增加 OrderKey OrderDesc CreatedAtBegin  CreatedAtEnd
        in: query
        name: orderKey
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: style
        type: string
      - in: query
        name: width
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取CmsAdSeat列表
      tags:
      - CmsAdSeat
  /cmsAdSeat/quickEdit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 快速更新CmsAdSeat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsAdSeat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 快速更新
      tags:
      - QuickEdit
  /cmsAdSeat/updateCmsAdSeat:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新CmsAdSeat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsAdSeat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新CmsAdSeat
      tags:
      - CmsAdSeat
...
  /cmsCat/createCmsCat:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建CmsCat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsCat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建CmsCat
      tags:
      - CmsCat
  /cmsCat/deleteCmsCat:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除CmsCat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsCat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除CmsCat
      tags:
      - CmsCat
  /cmsCat/deleteCmsCatByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除CmsCat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除CmsCat
      tags:
      - CmsCat
  /cmsCat/findCmsCat:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: alias
        type: string
      - in: query
        name: beNav
        type: boolean
      - in: query
        name: beSys
        type: boolean
      - in: query
        name: desc
        type: string
      - in: query
        name: groupId
        type: integer
      - in: query
        name: keywords
        type: string
      - in: query
        name: mediaType
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: pid
        type: integer
      - in: query
        name: sort
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: thumb
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询CmsCat
      tags:
      - CmsCat
  /cmsCat/getCmsCatList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: alias
        type: string
      - in: query
        name: beNav
        type: boolean
      - in: query
        name: beSys
        type: boolean
      - in: query
        name: desc
        type: string
      - in: query
        name: groupId
        type: integer
      - in: query
        name: keywords
        type: string
      - in: query
        name: mediaType
        type: integer
      - in: query
        name: name
        type: string
      - description: 排序方式:升序false(默认)|降序true
        in: query
        name: orderDesc
        type: boolean
      - description: Add By ljd 20210708, 统一在这里增加 OrderKey OrderDesc CreatedAtBegin  CreatedAtEnd
        in: query
        name: orderKey
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: pid
        type: integer
      - in: query
        name: sort
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: thumb
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取CmsCat列表
      tags:
      - CmsCat
  /cmsCat/quickEdit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 快速更新CmsCat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsCat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 快速更新
      tags:
      - QuickEdit
  /cmsCat/updateCmsCat:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新CmsCat
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.CmsCat'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新CmsCat
      tags:
      - CmsCat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"